# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppStoreCat API",
        "description": "Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:7460/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/publishers/search": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Search publishers across stores",
                "operationId": "searchPublishers",
                "parameters": [
                    {
                        "name": "term",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    },
                    {
                        "name": "platform",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "us",
                            "maxLength": 2,
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Publisher search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PublisherSearchResultResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/publishers": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "List publishers from user apps",
                "operationId": "listPublishers",
                "responses": {
                    "200": {
                        "description": "Publisher list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PublisherResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/publishers/{platform}/{externalId}": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Get publisher details with tracked apps",
                "operationId": "showPublisher",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Publisher details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublisherDetailResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/publishers/{platform}/{externalId}/store-apps": {
            "get": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Fetch all apps by publisher from store",
                "operationId": "publisherStoreApps",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Apps from store",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "apps": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/StoreAppResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/publishers/{platform}/{externalId}/import": {
            "post": {
                "tags": [
                    "Publishers"
                ],
                "summary": "Import and track apps from publisher",
                "operationId": "importPublisherApps",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "external_ids"
                                ],
                                "properties": {
                                    "external_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Apps imported"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Platform": {
                "type": "string",
                "example": "ios",
                "enum": [
                    "Ios",
                    "Android"
                ]
            },
...
            "AppResource": {
                "required": [
                    "id",
                    "name",
                    "platform",
                    "external_id",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Instagram"
                    },
                    "platform": {
                        "$ref": "#/components/schemas/Platform"
                    },
                    "external_id": {
                        "type": "string",
                        "example": "389801252"
                    },
                    "publisher": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Publisher"
                            }
                        ],
                        "nullable": true
                    },
                    "category": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/StoreCategory"
                            }
                        ],
                        "nullable": true
                    },
                    "icon_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.68,
                        "nullable": true
                    },
                    "rating_count": {
                        "type": "integer",
                        "example": 31,
                        "nullable": true
                    },
                    "version": {
                        "type": "string",
                        "example": "1.2.4",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PublisherDetailResource": {
                "properties": {
                    "publisher": {
                        "$ref": "#/components/schemas/PublisherResource"
                    },
                    "apps": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AppResource"
                        }
                    }
                },
                "type": "object"
            },
            "PublisherResource": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Publisher"
                    },
                    {
                        "properties": {
                            "apps_count": {
                                "type": "integer",
                                "example": 5,
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "PublisherSearchResultResource": {
                "required": [
                    "external_id",
                    "name",
                    "platform"
                ],
                "properties": {
                    "external_id": {
                        "type": "string",
                        "example": "389801255"
                    },
                    "name": {
                        "type": "string",
                        "example": "Meta Platforms, Inc."
                    },
                    "url": {
                        "type": "string",
                        "nullable": true
                    },
                    "platform": {
                        "type": "string",
                        "example": "ios"
                    },
                    "app_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "sample_apps": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "icon_url": {
                                    "type": "string",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "StoreAppResource": {
                "required": [
                    "external_id",
                    "name"
                ],
                "properties": {
                    "external_id": {
                        "type": "string",
                        "example": "389801252"
                    },
                    "name": {
                        "type": "string",
                        "example": "Instagram"
                    },
                    "icon_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "rating_count": {
                        "type": "integer",
                        "nullable": true
                    },
                    "is_free": {
                        "type": "boolean"
                    },
                    "category": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_tracked": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "Publisher": {
                "required": [
                    "id",
                    "name",
                    "platform"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Instagram, Inc."
                    },
                    "external_id": {
                        "type": "string",
                        "example": "389801255",
                        "nullable": true
                    },
                    "platform": {
                        "$ref": "#/components/schemas/Platform"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://apps.apple.com/us/developer/instagram-inc/id389801255",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreCategory": {
                "required": [
                    "id",
                    "name",
                    "slug",
                    "platform"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Photo & Video"
                    },
                    "slug": {
                        "type": "string",
                        "example": "photo-video"
                    },
                    "platform": {
                        "$ref": "#/components/schemas/Platform"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentication using Bearer token. Example: Bearer {your-token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}